// src/api/generated/models/diseaseDetailsResponseSchema.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * AI Agro API
 * API для AI Agro: регистрация, авторизация, создание чата и диагностика растений.
 * OpenAPI spec version: 1.0.0
 */
import type { DiseaseDetailSchema } from './diseaseDetailSchema';

export interface DiseaseDetailsResponseSchema {
  disease_name: string;
  sources: DiseaseDetailSchema[];
}